+++ src/ace_time/TimeZone.h
@@ -9,7 +9,8 @@
 #include <stdint.h>
 #include "TimeOffset.h"
 #include "ZoneProcessor.h"
-#include "ZoneProcessorCache.h"
+#include "BasicZoneProcessor.h"
+#include "ExtendedZoneProcessor.h"
 #include "BasicZone.h"
 #include "ExtendedZone.h"
 #include "TimeZoneData.h"
@@ -17,11 +18,6 @@
 class Print;
 
 namespace ace_time {
-
-// Allow ZoneManager to access the TimeZone() constructor that accepts
-// a ZoneProcessorCache.
-template<typename ZI, typename ZRR, typename ZPC>
-class ZoneManagerImpl;
 
 /**
  * Class that describes a time zone. There are 2 colloquial usages of "time
@@ -32,20 +28,16 @@ class ZoneManagerImpl;
  * source of these geographical regions is the TZ Database maintained by IANA
  * (https://www.iana.org/time-zones). The TimeZone class supports both meanings.
  *
- * There are 6 types of TimeZone:
+ * There are 4 types of TimeZone:
  *
  *    * kTypeError: represents an error or unknown time zone
  *    * kTypeManual: holds a base offset and a DST offset, and
  *      allows the user to modify both of these fields
  *    * kTypeBasic: using an underlying BasicZoneProcessor which
- *      supports 231 geographical zones in the TZ Database.
+ *      supports subset of zones and links defined by `zonedb/zone_infos.h`
  *    * kTypeExtended: a time zone using an underlying ExtendedZoneProcessor
- *      which supports 348 geographical zones in the TZ Database (essentially
- *      the entire database).
- *    * kTypeBasicManaged: created through the ZoneManager which contains
- *      an internal cache of BasicZoneProcessors.
- *    * kTypeExtendedManaged: created through the ZoneManager which contains
- *      an internal cache of ExtendedZoneProcessors.
+ *      which supports all zones and links defined by `zonedbx/zone_infos.h`,
+ *      essentially the entire IANA TZ database
  *
  * The TimeZone class should be treated as a const value type. (Except for
  * kTypeManual which is self-contained and allows the stdOffset and dstOffset
@@ -88,10 +80,6 @@ class TimeZone {
     static const uint8_t kTypeManual = 1;
     static const uint8_t kTypeBasic = ZoneProcessor::kTypeBasic;
     static const uint8_t kTypeExtended = ZoneProcessor::kTypeExtended;
-    static const uint8_t kTypeBasicManaged =
-        ZoneProcessorCache::kTypeBasicManaged;
-    static const uint8_t kTypeExtendedManaged =
-        ZoneProcessorCache::kTypeExtendedManaged;
 
     /** Factory method to create a UTC TimeZone. */
     static TimeZone forUtc() {
@@ -223,20 +211,17 @@ class TimeZone {
     }
 
     /**
-     * Return the zoneId for kTypeBasic, kTypeExtended, kTypeBasicManaged,
-     * kTypeExtendedManaged. Returns 0 for kTypeManual. (It is not entirely
-     * clear that a valid zoneId is always > 0, but there is little I can do
-     * without C++ exceptions.)
+     * Return the zoneId for kTypeBasic, kTypeExtended. Returns 0 for
+     * kTypeManual. (It is not entirely clear that a valid zoneId is always >
+     * 0, but there is little I can do without C++ exceptions.)
      */
     uint32_t getZoneId() const {
       switch (mType) {
         case kTypeManual:
           return 0;
         case kTypeBasic:
-        case kTypeBasicManaged:
           return BasicZone((const basic::ZoneInfo*) mZoneInfo).zoneId();
         case kTypeExtended:
-        case kTypeExtendedManaged:
           return ExtendedZone((const extended::ZoneInfo*) mZoneInfo).zoneId();
       }
       return 0;
@@ -256,14 +241,6 @@ class TimeZone {
         case kTypeExtended:
           mZoneProcessor->setZoneInfo(mZoneInfo);
           return mZoneProcessor->getUtcOffset(epochSeconds);
-        case kTypeBasicManaged:
-        case kTypeExtendedManaged:
-        {
-          ZoneProcessor* processor =
-              mZoneProcessorCache->getZoneProcessor(mZoneInfo);
-          if (! processor) break;
-          return processor->getUtcOffset(epochSeconds);
-        }
       }
       return TimeOffset::forError();
     }
@@ -281,14 +258,6 @@ class TimeZone {
         case kTypeExtended:
           mZoneProcessor->setZoneInfo(mZoneInfo);
           return mZoneProcessor->getDeltaOffset(epochSeconds);
-        case kTypeBasicManaged:
-        case kTypeExtendedManaged:
-        {
-          ZoneProcessor* processor =
-              mZoneProcessorCache->getZoneProcessor(mZoneInfo);
-          if (! processor) break;
-          return processor->getDeltaOffset(epochSeconds);
-        }
       }
       return TimeOffset::forError();
     }
@@ -296,10 +265,9 @@ class TimeZone {
     /**
      * Return the time zone abbreviation at the given epochSeconds.
      *
-     *   * kTypeManual returns "STD" or "DST",
-     *   * kTypeBasic, kTypeBasic, kTypeBasicManaged, kTypeExtendedManaged
-     *      * returns the "{abbrev}" (e.g. "PDT"),
-     *   * the empty string "" upon error.
+     *   * kTypeManual: returns "STD" or "DST",
+     *   * kTypeBasic, kTypeBasic: returns the "{abbrev}" (e.g. "PDT"),
+     *   * error condition: returns the empty string ""
      *
      * The IANA spec
      * (https://data.iana.org/time-zones/theory.html#abbreviations) says that
@@ -324,14 +292,6 @@ class TimeZone {
         case kTypeBasic:
         case kTypeExtended:
           return mZoneProcessor->getAbbrev(epochSeconds);
-        case kTypeBasicManaged:
-        case kTypeExtendedManaged:
-        {
-          ZoneProcessor* processor =
-              mZoneProcessorCache->getZoneProcessor(mZoneInfo);
-          if (! processor) break;
-          return processor->getAbbrev(epochSeconds);
-        }
       }
       return "";
     }
@@ -354,15 +314,6 @@ class TimeZone {
           mZoneProcessor->setZoneInfo(mZoneInfo);
           odt = mZoneProcessor->getOffsetDateTime(ldt);
           break;
-        case kTypeBasicManaged:
-        case kTypeExtendedManaged:
-        {
-          ZoneProcessor* processor =
-              mZoneProcessorCache->getZoneProcessor(mZoneInfo);
-          if (! processor) break;
-          odt = processor->getOffsetDateTime(ldt);
-          break;
-        }
       }
       return odt;
     }
@@ -405,9 +356,9 @@ class TimeZone {
 
     /**
      * Convert to a TimeZoneData object, which can be fed back into
-     * ZoneManager::createForTimeZoneData() to recreate the TimeZone. All of
-     * TimeZone::kTypeBasic, kTypeExtended, kTypeBasicManaged,
-     * kTypeExtendedManaged collapse into TimeZoneData::kTypeZoneId.
+     * ZoneManager::createForTimeZoneData() to recreate the TimeZone. Both
+     * TimeZone::kTypeBasic and TimeZone::kTypeExtended are mapped to
+     * TimeZoneData::kTypeZoneId.
      */
     TimeZoneData toTimeZoneData() const {
       TimeZoneData d;
@@ -419,8 +370,6 @@ class TimeZone {
           break;
         case TimeZone::kTypeBasic:
         case TimeZone::kTypeExtended:
-        case TimeZone::kTypeBasicManaged:
-        case TimeZone::kTypeExtendedManaged:
           d.zoneId = getZoneId();
           d.type = TimeZoneData::kTypeZoneId;
           break;
@@ -455,27 +404,6 @@ class TimeZone {
   private:
     friend bool operator==(const TimeZone& a, const TimeZone& b);
 
-    // Allow ZoneManager to access the TimeZone() constructor that accepts
-    // a ZoneProcessorCache.
-    template<typename ZI, typename ZRR, typename ZPC>
-    friend class ZoneManagerImpl;
-
-    /**
-     * Constructor for kType*Managed. Intended to be used ONLY by
-     * BasicZoneManager and ExtendedZoneManager.
-     *
-     * @param zoneInfo a pointer to a basic::ZoneInfo or an extended::ZoneInfo.
-     *        Cannot be nullptr.
-     * @param zoneProcessorCache a cache of ZoneInfo -> ZoneProcessor
-     */
-    explicit TimeZone(
-        const void* zoneInfo,
-        ZoneProcessorCache* zoneProcessorCache
-    ) :
-        mType(zoneProcessorCache->getType()),
-        mZoneInfo(zoneInfo),
-        mZoneProcessorCache(zoneProcessorCache) {}
-
     /**
      * Constructor for a kTypeManual TimeZone.
      *
@@ -500,11 +428,10 @@ class TimeZone {
 
     uint8_t mType;
 
-    // 4 combinations:
+    // 3 combinations:
     //   (type) (kTypeError)
     //   (type, mStdOffsetMinutes, mDstOffsetMinutes)
     //   (type, mZoneInfo, mZoneProcessor)
-    //   (type, mZoneInfo, mZoneProcessorCache)
     union {
       /** Used by kTypeManual. */
       struct {
@@ -512,20 +439,13 @@ class TimeZone {
         int16_t mDstOffsetMinutes;
       };
 
+      /* Used by kTypeBasic and kTypeExtended. */
       struct {
-        /**
-         * Either a basic::ZoneInfo or an extended::ZoneInfo. Used by
-         * kTypeBasic, kTypeExtended, kTypeBasicManaged, kTypeExtendedManaged.
-         */
+        /** Either a basic::ZoneInfo or an extended::ZoneInfo. */
         const void* mZoneInfo;
 
-        union {
-          /** Used by kTypeBasic, kTypeExtended. */
-          ZoneProcessor* mZoneProcessor;
-
-          /** Used by kTypeBasicManaged, kTypeExtendedManaged. */
-          ZoneProcessorCache* mZoneProcessorCache;
-        };
+        /** Either BasicZoneProcessor or ExtendedZoneProcessor. */
+        ZoneProcessor* mZoneProcessor;
       };
     };
 };
@@ -540,8 +460,6 @@ inline bool operator==(const TimeZone& a, const TimeZone& b) {
           && a.mDstOffsetMinutes == b.mDstOffsetMinutes;
     case TimeZone::kTypeBasic:
     case TimeZone::kTypeExtended:
-    case TimeZone::kTypeBasicManaged:
-    case TimeZone::kTypeExtendedManaged:
       return (a.mZoneInfo == b.mZoneInfo);
     default:
       return false;